requirements.txt
clint
colorama
mysql-connector-python
prettytable
pyautogui
art
pynput
python-dotenv
PyYAML
loguru
black
rich
setproctitle
psutil
watchdog